requirements.txt
asgiref==3.8.1
Django==4.2.19
django-bootstrap5==22.2
django_debug_toolbar==3.8.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
